make a reusable hook
