+++ components/pages/explorer/ResizeSidebar.tsx
@@ -1,11 +1,32 @@
-import { Chevron } from '@/components/theme/icons';
-import { css, useTheme } from '@emotion/react';
-import { ReactNode } from 'react';
-import { Rnd } from 'react-rnd';
+/*
+ *
+ * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ *  This program and the accompanying materials are made available under the terms of
+ *  the GNU Affero General Public License v3.0. You should have received a copy of the
+ *  GNU Affero General Public License along with this program.
+ *   If not, see <http://www.gnu.org/licenses/>.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ *  EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ *  OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ *  SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ *  TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ *  OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ *  IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ */
 
-export const sidebarHeight = '100%';
-export const sidebarToggleWidth = 28;
-export const sidebarMinWidth = 150;
+import { ReactNode, useEffect, useState } from 'react';
+import { Rnd } from 'react-rnd';
+import { css, useTheme } from '@emotion/react';
+import { Chevron } from '@/components/theme/icons';
+
+const sidebarHeight = '100%';
+const sidebarToggleWidth = 28;
+const sidebarMinWidth = 150;
 
 export const SidebarResizeWrapper = ({
 	children,
@@ -113,4 +134,39 @@ export const SidebarToggle = ({
 			</button>
 		</div>
 	);
+};
+
+export const useResizeSidebar = () => {
+	const theme = useTheme();
+	const sidebarDefaultWidth = theme.dimensions.facets.width || 0;
+	const [sidebarWidth, setSidebarWidth] = useState<number>(sidebarDefaultWidth);
+	const [sidebarVisible, setSidebarVisible] = useState<boolean>(true);
+
+	const toggleSidebarVisible = () => {
+		// toggle visibility with CSS display block/none
+		// rather than a JSX ternary statement
+		// in order to maintain the state of the facet panel
+		setSidebarVisible(!sidebarVisible);
+		setSidebarWidth(sidebarDefaultWidth);
+	};
+
+	const contentOffset = sidebarVisible ? sidebarWidth : sidebarToggleWidth;
+
+	useEffect(() => {
+		// when the sidebar is resized to smaller than sidebarMinWidth,
+		// collapse the sidebar
+		if (sidebarWidth < sidebarMinWidth) {
+			setSidebarVisible(false);
+		}
+	}, [sidebarWidth]);
+
+	return {
+		contentOffset,
+		setSidebarVisible,
+		setSidebarWidth,
+		sidebarDefaultWidth,
+		sidebarVisible,
+		sidebarWidth,
+		toggleSidebarVisible,
+	};
 };
